MongoDB Connected to React Simple
Remade the file and then attached mongodb and react another way. Created a new cluster on mongodb and then inserted the recipe json.

Currently just run the server and then the client and it displays everything in the mongodb server onto the html page. It obv needs CSS and organizing still